chore: extract git operations to separate package
